Fix: floats for `_f()` functions
• Accidentally left in `int`s for `clamp_f()` and `clamp01_f()` args and return values— making them truncate the the float values by automatic downcast and re-upcast.  Whoops.

+++ DrawInQuad/CGTextureMapping.cpp
@@ -79,7 +79,7 @@ static inline int clamp_i(int v, int l, int h)
 	else
 		return v;
 }
-static inline int clamp_f(int v, int l, int h)
+static inline float clamp_f(float v, float l, float h)
 {
 	if (v < l)
 		return l;
@@ -88,7 +88,7 @@ static inline int clamp_f(int v, int l, int h)
 	else
 		return v;
 }
-static inline int clamp01_f(int v) {
+static inline float clamp01_f(float v) {
 	return clamp_f(v, 0.0f, 1.0f);
 }
